requirements/dev.txt
# Local development dependencies go here
-r base.txt

# Debug & Testing
django-debug-toolbar==1.3.0
sqlparse==0.1.14
flake8==2.4.0
factory_boy==2.5.1
